Refactored and moved some classes
