--- src/main/java/babinvas/insurancecontrolaccessdb/services/EmailSender.java
@@ -1,9 +1,0 @@
-package babinvas.insurancecontrolaccessdb.services;
-
-public interface EmailSender {
-	void setFrom(String from);
-	void setTo(String to);
-	void setCc(String cc);
-	void setBcc(String bcc);
-	void send();
-}
