+++ src/main/java/babinvas/insurancecontrolaccessdb/services/email/EmailSendingService.java
@@ -1,0 +1,9 @@
+package babinvas.insurancecontrolaccessdb.services.email;
+
+public interface EmailSendingService {
+	void setFrom(String from);
+	void setTo(String to);
+	void setCc(String cc);
+	void setBcc(String bcc);
+	void send(String subject, String text);
+}
